chore(ci): handle workflow error
