+++ src/Apps/GameDesktop/Factories/LogFactory.cs
