+++ src/Apps/GameDesktop/GameDesktop.csproj
@@ -27,13 +27,21 @@
   <ItemGroup>
     <EmbeddedResource Include="Icon.ico" />
     <EmbeddedResource Include="Icon.bmp" />
-    <EmbeddedResource Update="Resources\Configs.resx">
-      <Generator>ResXFileCodeGenerator</Generator>
-      <LastGenOutput>Configs.Designer.cs</LastGenOutput>
-    </EmbeddedResource>
     <EmbeddedResource Update="Resources\SpriteSheets.resx">
       <Generator>ResXFileCodeGenerator</Generator>
       <LastGenOutput>SpriteSheets.Designer.cs</LastGenOutput>
+    </EmbeddedResource>
+    <EmbeddedResource Update="Resources\Errors.resx">
+      <Generator>ResXFileCodeGenerator</Generator>
+      <LastGenOutput>Errors.Designer.cs</LastGenOutput>
+    </EmbeddedResource>
+    <EmbeddedResource Update="Resources\AppVariables.resx">
+      <Generator>ResXFileCodeGenerator</Generator>
+      <LastGenOutput>Env.Designer.cs</LastGenOutput>
+    </EmbeddedResource>
+    <EmbeddedResource Update="Resources\EnvironmentVariables.resx">
+      <Generator>ResXFileCodeGenerator</Generator>
+      <LastGenOutput>EnvironmentVariables.Designer.cs</LastGenOutput>
     </EmbeddedResource>
   </ItemGroup>
   <ItemGroup>
@@ -63,15 +71,25 @@
     <ProjectReference Include="..\..\Libs\Services\Services.csproj" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Update="Resources\Configs.Designer.cs">
-      <DesignTime>True</DesignTime>
-      <AutoGen>True</AutoGen>
-      <DependentUpon>Configs.resx</DependentUpon>
-    </Compile>
     <Compile Update="Resources\SpriteSheets.Designer.cs">
       <DesignTime>True</DesignTime>
       <AutoGen>True</AutoGen>
       <DependentUpon>SpriteSheets.resx</DependentUpon>
+    </Compile>
+    <Compile Update="Resources\Errors.Designer.cs">
+      <DesignTime>True</DesignTime>
+      <AutoGen>True</AutoGen>
+      <DependentUpon>Errors.resx</DependentUpon>
+    </Compile>
+    <Compile Update="Resources\AppVariables.Designer.cs">
+      <DesignTime>True</DesignTime>
+      <AutoGen>True</AutoGen>
+      <DependentUpon>AppVariables.resx</DependentUpon>
+    </Compile>
+    <Compile Update="Resources\EnvironmentVariables.Designer.cs">
+      <DesignTime>True</DesignTime>
+      <AutoGen>True</AutoGen>
+      <DependentUpon>EnvironmentVariables.resx</DependentUpon>
     </Compile>
   </ItemGroup>
   <Target Name="RestoreDotnetTools" BeforeTargets="Restore">
