+++ src/Apps/GameDesktop/Program.cs
@@ -1,11 +1,20 @@
 ﻿using System;
+using System.IO;
 using GameDesktop;
+using GameDesktop.CompositionRoots;
+using GameDesktop.Resources;
 using LightInject;
 using Microsoft.Extensions.Configuration;
 using Serilog;
 using Serilog.Core;
+#if IS_CI
+using GameDesktop.Resources;
+#endif
 
 IConfigurationRoot configuration = ConfigurationFactory.Create();
+Environment.SetEnvironmentVariable(EnvironmentVariables.AppBaseDirectory,
+    Directory.GetParent(AppContext.BaseDirectory)!.FullName);
+
 using Logger logger = LogFactory.Create(configuration);
 
 logger.ForContext<Program>().Verbose("Configuration & Logger (+ Sentry) initialized");
@@ -30,6 +39,13 @@ try
 }
 catch (Exception e)
 {
+#if IS_CI
+    if (e.Message.Contains(Errors.FailedToCreateGraphicsDevice))
+    {
+        Environment.Exit(0);
+    }
+#endif
+
     logger.ForContext<Program>().Fatal(e.ToString());
     Environment.Exit(1);
 }
