+++ src/Apps/GameDesktop/Resources/Errors.resx
@@ -1,4 +1,4 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
+<?xml version="1.0" encoding="utf-8"?>
 
 <root>
   <xsd:schema id="root" xmlns="" xmlns:xsd="http://www.w3.org/2001/XMLSchema"
@@ -23,4 +23,6 @@
       PublicKeyToken=b77a5c561934e089
     </value>
   </resheader>
+
+  <data name="FailedToCreateGraphicsDevice">Failed to create graphics device!</data>
 </root>
