+++ src/Apps/GameDesktop/appsettings.Development.json
@@ -1,2 +1,21 @@
 ﻿{
+  "Serilog": {
+    "Using": [
+      "Serilog.Sinks.Console",
+      "Serilog.Sinks.File"
+    ],
+    "MinimumLevel": "Verbose",
+    "Enrich": [
+      "FromLogContext"
+    ],
+    "WriteTo": [
+      {
+        "Name": "Console",
+        "Args": {
+          "outputTemplate": "[{Level:u3}] [{SourceContext}] {Message:lj} {NewLine}{Exception}",
+          "theme": "Serilog.Sinks.SystemConsole.Themes.SystemConsoleTheme::Colored, Serilog.Sinks.Console"
+        }
+      }
+    ]
+  }
 }
