+++ src/Apps/GameDesktop/appsettings.Production.json
@@ -1,2 +1,35 @@
 ﻿{
+  "Serilog": {
+    "Using": [
+      "Serilog.Sinks.Console",
+      "Serilog.Sinks.File"
+    ],
+    "MinimumLevel": "Verbose",
+    "Enrich": [
+      "FromLogContext"
+    ],
+    "WriteTo": [
+      {
+        "Name": "File",
+        "Args": {
+          "path": "%APP_BASE_DIRECTORY%/Logs/log-.txt",
+          "outputTemplate": "{Timestamp:yyyy-MM-dd HH:mm:ss.fff zzz} [{Level:u3}] [{SourceContext}] {Message:lj} {NewLine}{Exception}",
+          "rollingInterval": "Day"
+        }
+      },
+      {
+        "Name": "Sentry",
+        "Args": {
+          "EnableTracing": true,
+          "MaxRequestBodySize": "Always",
+          "SendDefaultPii": true,
+          "IncludeActivityData": true,
+          "AttachStackTrace": true,
+          "Debug": true,
+          "Dsn": "https://ff3f6fec4457d740ab0a98c123e77086@o4505883399487488.ingest.sentry.io/4505883401388032",
+          "DiagnosticLevel": "Error"
+        }
+      }
+    ]
+  }
 }
