+++ src/Apps/GameDesktop/appsettings.json
@@ -1,42 +1,2 @@
 ﻿{
-  "Serilog": {
-    "Using": [
-      "Serilog.Sinks.Console",
-      "Serilog.Sinks.File"
-    ],
-    "MinimumLevel": "Verbose",
-    "Enrich": [
-      "FromLogContext"
-    ],
-    "WriteTo": [
-      {
-        "Name": "Console",
-        "Args": {
-          "outputTemplate": "[{Level:u3}] [{SourceContext}] {Message:lj} {NewLine}{Exception}",
-          "theme": "Serilog.Sinks.SystemConsole.Themes.SystemConsoleTheme::Colored, Serilog.Sinks.Console"
-        }
-      },
-      {
-        "Name": "File",
-        "Args": {
-          "path": "%APPDATA%/GameDesktop/Logs/log-.txt",
-          "outputTemplate": "{Timestamp:yyyy-MM-dd HH:mm:ss.fff zzz} [{Level:u3}] [{SourceContext}] {Message:lj} {NewLine}{Exception}",
-          "rollingInterval": "Day"
-        }
-      },
-      {
-        "Name": "Sentry",
-        "Args": {
-          "EnableTracing": true,
-          "MaxRequestBodySize": "Always",
-          "SendDefaultPii": true,
-          "IncludeActivityData": true,
-          "AttachStackTrace": true,
-          "Debug": true,
-          "Dsn": "https://ff3f6fec4457d740ab0a98c123e77086@o4505883399487488.ingest.sentry.io/4505883401388032",
-          "DiagnosticLevel": "Error"
-        }
-      }
-    ]
-  }
 }
